Move get_entity, embed_references and other DAO specific methods to dao.utils
