+++ metadata_repository_service/dao/data_access_committee.py
@@ -19,14 +19,10 @@ Convenience methods for retrieving DataAccessCommittee records
 from typing import List
 
 from metadata_repository_service.config import CONFIG, Config
-from metadata_repository_service.core.utils import (
-    generate_accession,
-    generate_uuid,
-    get_entity,
-    get_timestamp,
-)
+from metadata_repository_service.core.utils import generate_uuid, get_timestamp
 from metadata_repository_service.dao.db import get_db_client
 from metadata_repository_service.dao.member import create_member, get_member_by_email
+from metadata_repository_service.dao.utils import generate_accession, get_entity
 from metadata_repository_service.models import (
     CreateDataAccessCommittee,
     DataAccessCommittee,
