+++ metadata_repository_service/dao/data_access_policy.py
@@ -19,16 +19,12 @@ Convenience methods for retrieving DataAccessPolicy records
 from typing import List
 
 from metadata_repository_service.config import CONFIG, Config
-from metadata_repository_service.core.utils import (
-    generate_accession,
-    generate_uuid,
-    get_entity,
-    get_timestamp,
-)
+from metadata_repository_service.core.utils import generate_uuid, get_timestamp
 from metadata_repository_service.dao.data_access_committee import (
     get_data_access_committee_by_accession,
 )
 from metadata_repository_service.dao.db import get_db_client
+from metadata_repository_service.dao.utils import generate_accession, get_entity
 from metadata_repository_service.models import CreateDataAccessPolicy, DataAccessPolicy
 
 COLLECTION_NAME = "DataAccessPolicy"
