+++ metadata_repository_service/dao/member.py
@@ -19,12 +19,9 @@ Convenience methods for retrieving Member records
 from typing import List
 
 from metadata_repository_service.config import CONFIG, Config
-from metadata_repository_service.core.utils import (
-    generate_uuid,
-    get_entity,
-    get_timestamp,
-)
+from metadata_repository_service.core.utils import generate_uuid, get_timestamp
 from metadata_repository_service.dao.db import get_db_client
+from metadata_repository_service.dao.utils import get_entity
 from metadata_repository_service.models import CreateMember, Member
 
 COLLECTION_NAME = "Member"
